try to create CurrencyClient

+++ product-api/data/products.go
@@ -2,11 +2,11 @@ package data
 
 // Product is entities
 type Product struct {
-	ID          int    `json:"id"`
-	Name        string `json:"name" validate:"required"`
-	Description string `json:"description"`
-	Price       int32  `json:"price" validate:"gt=0"`
-	SKU         string `json:"sku" validate:"required,sku"`
+	ID          int     `json:"id"`
+	Name        string  `json:"name" validate:"required"`
+	Description string  `json:"description"`
+	Price       float32 `json:"price" validate:"gt=0"`
+	SKU         string  `json:"sku" validate:"required,sku"`
 }
 
 // Products is a collection of Product
